+++ frontend/src/app/pages/exercises/exercise/shared/utils/types/positioning.ts
